refactor: numpy utils
