+++ queens/parameters/random_fields/random_field.py
@@ -18,7 +18,7 @@ import abc
 
 import numpy as np
 
-from queens.utils.numpy import at_least_2d
+from queens.utils.numpy_array import at_least_2d
 
 
 class RandomField(metaclass=abc.ABCMeta):
